# docs/api/openapi.json
{
  "components": {
    "schemas": {
...
      "CalendarConfigIn": {
        "properties": {
          "effective_from": {
            "format": "date",
            "title": "Effective From",
            "type": "string"
          },
          "grace_minutes": {
            "default": 0,
            "maximum": 1440.0,
            "minimum": 0.0,
            "title": "Grace Minutes",
            "type": "integer"
          },
          "meeting_weekday": {
            "default": 0,
            "maximum": 6.0,
            "minimum": 0.0,
            "title": "Meeting Weekday",
            "type": "integer"
          },
          "timezone": {
            "default": "Asia/Ho_Chi_Minh",
            "title": "Timezone",
            "type": "string"
          },
          "week_start_weekday": {
            "default": 0,
            "maximum": 6.0,
            "minimum": 0.0,
            "title": "Week Start Weekday",
            "type": "integer"
          }
        },
        "required": [
          "effective_from"
        ],
        "title": "CalendarConfigIn",
        "type": "object"
      },
      "CalendarConfigOut": {
        "properties": {
          "effective_from": {
            "format": "date",
            "title": "Effective From",
            "type": "string"
          },
          "grace_minutes": {
            "title": "Grace Minutes",
            "type": "integer"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "meeting_weekday": {
            "title": "Meeting Weekday",
            "type": "integer"
          },
          "timezone": {
            "title": "Timezone",
            "type": "string"
          },
          "version": {
            "title": "Version",
            "type": "integer"
          },
          "week_start_weekday": {
            "title": "Week Start Weekday",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "version",
          "timezone",
          "meeting_weekday",
          "week_start_weekday",
          "grace_minutes",
          "effective_from"
        ],
        "title": "CalendarConfigOut",
        "type": "object"
      },
...
      "DraftIn": {
        "properties": {
          "content": {
            "additionalProperties": true,
            "title": "Content",
            "type": "object"
          }
        },
        "title": "DraftIn",
        "type": "object"
      },
...
      "ExcuseIn": {
        "properties": {
          "reason": {
            "title": "Reason",
            "type": "string"
          }
        },
        "required": [
          "reason"
        ],
        "title": "ExcuseIn",
        "type": "object"
      },
      "ExtensionIn": {
        "properties": {
          "reason": {
            "title": "Reason",
            "type": "string"
          },
          "until": {
            "format": "date-time",
            "title": "Until",
            "type": "string"
          }
        },
        "required": [
          "until",
          "reason"
        ],
        "title": "ExtensionIn",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ObligationOut": {
        "properties": {
          "excuse_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Excuse Reason"
          },
          "extension_until_utc": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Extension Until Utc"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "membership_id": {
            "format": "uuid",
            "title": "Membership Id",
            "type": "string"
          },
          "period_id": {
            "format": "uuid",
            "title": "Period Id",
            "type": "string"
          },
          "project_id": {
            "format": "uuid",
            "title": "Project Id",
            "type": "string"
          },
          "state": {
            "$ref": "#/components/schemas/ObligationState"
          },
          "student_id": {
            "format": "uuid",
            "title": "Student Id",
            "type": "string"
          },
          "submitted": {
            "default": false,
            "title": "Submitted",
            "type": "boolean"
          }
        },
        "required": [
          "id",
          "period_id",
          "membership_id",
          "student_id",
          "project_id",
          "state"
        ],
        "title": "ObligationOut",
        "type": "object"
      },
      "ObligationState": {
        "enum": [
          "required",
          "excused"
        ],
        "title": "ObligationState",
        "type": "string"
      },
...
      "PeriodOut": {
        "properties": {
          "deadline_utc": {
            "format": "date-time",
            "title": "Deadline Utc",
            "type": "string"
          },
          "end_utc": {
            "format": "date-time",
            "title": "End Utc",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "local_end": {
            "format": "date",
            "title": "Local End",
            "type": "string"
          },
          "local_start": {
            "format": "date",
            "title": "Local Start",
            "type": "string"
          },
          "meeting_date": {
            "format": "date",
            "title": "Meeting Date",
            "type": "string"
          },
          "reminder_due_utc": {
            "format": "date-time",
            "title": "Reminder Due Utc",
            "type": "string"
          },
          "start_utc": {
            "format": "date-time",
            "title": "Start Utc",
            "type": "string"
          },
          "workspace_id": {
            "format": "uuid",
            "title": "Workspace Id",
            "type": "string"
          }
        },
        "required": [
          "id",
          "workspace_id",
          "local_start",
          "local_end",
          "start_utc",
          "end_utc",
          "meeting_date",
          "deadline_utc",
          "reminder_due_utc"
        ],
        "title": "PeriodOut",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Research Management System",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/calendar": {
      "get": {
        "description": "`null` when no calendar has been configured, which is a state and not an error.\n\nReadable by a student as well as a professor: `CalendarConfig`'s policy is workspace-wide\nbecause everyone needs to know when their report is due (REP-01).",
        "operationId": "current_calendar_api_v1_calendar_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/CalendarConfigOut"
                    },
                    {
                      "type": "null"
                    }
                  ],
                  "title": "Response Current Calendar Api V1 Calendar Get"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "The reporting calendar in force",
        "tags": [
          "reporting"
        ]
      },
      "put": {
        "operationId": "configure_calendar_api_v1_calendar_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CalendarConfigIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CalendarConfigOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Configure the reporting calendar",
        "tags": [
          "reporting"
        ]
      }
    },
...
    "/api/v1/obligations/{obligation_id}/excuse": {
      "post": {
        "operationId": "excuse_obligation_api_v1_obligations__obligation_id__excuse_post",
        "parameters": [
          {
            "in": "path",
            "name": "obligation_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Obligation Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExcuseIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ObligationOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Excuse an obligation",
        "tags": [
          "reporting"
        ]
      }
    },
    "/api/v1/obligations/{obligation_id}/extend": {
      "post": {
        "operationId": "extend_obligation_api_v1_obligations__obligation_id__extend_post",
        "parameters": [
          {
            "in": "path",
            "name": "obligation_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Obligation Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExtensionIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ObligationOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Extend one obligation's deadline",
        "tags": [
          "reporting"
        ]
      }
    },
...
    "/api/v1/periods": {
      "get": {
        "description": "The weeks of the workspace being worked in.",
        "operationId": "list_periods_api_v1_periods_get",
        "parameters": [
          {
            "in": "query",
            "name": "through",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Through"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/PeriodOut"
                  },
                  "title": "Response List Periods Api V1 Periods Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List reporting periods",
        "tags": [
          "reporting"
        ]
      }
    },
    "/api/v1/periods/ensure": {
      "post": {
        "operationId": "ensure_periods_api_v1_periods_ensure_post",
        "parameters": [
          {
            "in": "query",
            "name": "through",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Through"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/PeriodOut"
                  },
                  "title": "Response Ensure Periods Api V1 Periods Ensure Post",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Materialise periods up to a date",
        "tags": [
          "reporting"
        ]
      }
    },
    "/api/v1/periods/{period_id}/obligations": {
      "get": {
        "operationId": "list_obligations_api_v1_periods__period_id__obligations_get",
        "parameters": [
          {
            "in": "path",
            "name": "period_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Period Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "student_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Student Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ObligationOut"
                  },
                  "title": "Response List Obligations Api V1 Periods  Period Id  Obligations Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Who owes a report this period",
        "tags": [
          "reporting"
        ]
      }
    },
    "/api/v1/periods/{period_id}/obligations/ensure": {
      "post": {
        "operationId": "ensure_obligations_api_v1_periods__period_id__obligations_ensure_post",
        "parameters": [
          {
            "in": "path",
            "name": "period_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Period Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ObligationOut"
                  },
                  "title": "Response Ensure Obligations Api V1 Periods  Period Id  Obligations Ensure Post",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Derive obligations",
        "tags": [
          "reporting"
        ]
      }
    },
    "/api/v1/periods/{period_id}/report": {
      "get": {
        "operationId": "get_report_api_v1_periods__period_id__report_get",
        "parameters": [
          {
            "in": "path",
            "name": "period_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Period Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "student_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Student Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReportOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Read a weekly report",
        "tags": [
          "reporting"
        ]
      }
    },
    "/api/v1/periods/{period_id}/report/draft": {
      "patch": {
        "operationId": "save_draft_api_v1_periods__period_id__report_draft_patch",
        "parameters": [
          {
            "in": "path",
            "name": "period_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Period Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DraftIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReportOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Autosave the weekly draft",
        "tags": [
          "reporting"
        ]
      }
    },
    "/api/v1/periods/{period_id}/report/submit": {
      "post": {
        "description": "A repeated submission with the same Idempotency-Key returns the version already written.",
        "operationId": "submit_report_api_v1_periods__period_id__report_submit_post",
        "parameters": [
          {
            "in": "path",
            "name": "period_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Period Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 128,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SubmitIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VersionOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Submit the weekly package",
        "tags": [
          "reporting"
        ]
      }
    },
...
    "/api/v1/report-versions/{version_id}": {
      "get": {
        "operationId": "get_version_api_v1_report_versions__version_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "version_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Version Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VersionOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Read one submitted version",
        "tags": [
          "reporting"
        ]
      }
    },
    "/api/v1/reports/{report_id}/reviewed": {
      "post": {
        "operationId": "mark_reviewed_api_v1_reports__report_id__reviewed_post",
        "parameters": [
          {
            "in": "path",
            "name": "report_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Report Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReportOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Mark a report reviewed",
        "tags": [
          "reporting"
        ]
      }
    },
    "/api/v1/reports/{report_id}/revisions": {
      "get": {
        "operationId": "list_revision_requests_api_v1_reports__report_id__revisions_get",
        "parameters": [
          {
            "in": "path",
            "name": "report_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Report Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/RevisionRequestOut"
                  },
                  "title": "Response List Revision Requests Api V1 Reports  Report Id  Revisions Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Outstanding revision requests",
        "tags": [
          "reporting"
        ]
      },
      "post": {
        "operationId": "request_revision_api_v1_reports__report_id__revisions_post",
        "parameters": [
          {
            "in": "path",
            "name": "report_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Report Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RevisionRequestIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RevisionRequestOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Request a revision of one project entry",
        "tags": [
          "reporting"
        ]
      }
    },
    "/api/v1/reports/{report_id}/versions": {
      "get": {
        "description": "REP-05: a resubmission adds a version and never replaces history \u2014 so history needs a reader.\n\n`ScopeDep`, not `ProfScopeDep`: a student is entitled to their own, and the visibility\npredicate already confines them to it.",
        "operationId": "list_versions_api_v1_reports__report_id__versions_get",
        "parameters": [
          {
            "in": "path",
            "name": "report_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Report Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/VersionSummaryOut"
                  },
                  "title": "Response List Versions Api V1 Reports  Report Id  Versions Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Every submitted version of one report",
        "tags": [
          "reporting"
        ]
      }
    },
...
  }
}